AM57X-63 thermal: add support for two_stage thermal governor
This governor was created to explicitly enable and disable
a cooling device with only two stages as long as the
temperature is still above the trip_zone.

The step_wise governor's trend-based throttling resulted in
frequent, abrupt off and on toggling for a two-stage device
and this governor alleviates that behavior.

Signed-off-by: Russell Robinson <rrobinson@phytec.com>