AM57X-83 arm: dts: am57xx-phycore-rdk: update and enable lcd-018
Update lcd-018 nodes to properly use vout2, along with
improved hardware mux timings to reduce flickering.

The VOUT2 signals are part of two different PADCONF/MUX groups that
do not have predefined IODELAY timing values to align them so the
flickering is still present even with these improved timings, but
greatly diminished.

Also set the LCD as the default display and disable HDMI. HDMI is
capable of working in parallel, but the current filesystem loads
the TI Matrix Demo, which is not capable of utilizing both displays
simultaneously and results in varied behavior.

Signed-off-by: Russell Robinson <rrobinson@phytec.com>