AM57X-131 HACK: video: fbdev: omap2: dss: enable VOUT2
TI has not implemented VOUT2 for the AM57xx platform.

This is a hack for supporting VOUT2 on our RDK. The PLLs
were not being assigned to VOUT2, resulting in omapdrm errors
and a failure to display anything but a generic color test.

With this hack in place, there is a known issue of
overclocking/syncing on our display that will be resolved
in future releases.

Signed-off-by: Russell Robinson <rrobinson@phytec.com>