arm: dts: am57xx-pcm-948: fix broken SATA configuration
The AHCI driver was updated to assume firmware programs
the PORTS_IMPL register, but not all SOCs do this, meaning
the default number of ports (0) are available for use.

The ports-implemented dts binding was added with
	commit id 6e337a05df8adfc54540ca2a2b9d621836697796
as a workaround for this register not being programmed.

ports-implemented = <1>; makes 1 SATA port available.

Signed-off-by: Russell Robinson <rrobinson@phytec.com>