AM57X-131 gpu: drm: omapdrm: add support for VOUT2
Despite many of the functions being channel agnostic and generic,
the important functions for AM57xx use are hard-coded to only use
VOUT1 and the video0 PLL for all display output. This patch adds
support for VOUT2 to the current infrastructure.

In order to do that, the dispc channel is now aligned with the
proper clkout setting, and the driver more intelligently configures
VOUT1 to use the video0 PLL and VOUT2 the video1 PLL. Additionally,
the PLL values for clkout2 and clkout3 were previously missing from
the video-pll file and are now accessible

Signed-off-by: Russell Robinson <rrobinson@phytec.com>