TIAM57X-291 arm: dts: am57xx-pcm-948: fix USB2 vbus detection
When USB2 was used in host-mode and a gadget driver was loaded,
using the poweroff command would power off the board, but it would
immediately reboot due to VBUS not being shut off completely and
restarting the PMIC POWER_ON process.

With the VBUSDET signal out of the PMIC being handled properly
everything works as expected.

Signed-off-by: Russell Robinson <rrobinson@phytec.com>