AM57TLY-133 arm: dts: am57xx-phycore-som: fix the SPI NOR flash detection
QSPI1_D2 and QSPI1_D3 are not pulled up externally and
the pin-muxing was initially changed in support for SPI-QIO
mode, but due to silicon errata
	i916: QSPI Reads Can Fail For Flash Devices with HOLD Function
we configured the bus for SPI-DIO mode.

In the process, the pin-muxing was not reconfigured to pull
the two signals up internally so nHOLD and nWP were left floating.

Signed-off-by: Russell Robinson <rrobinson@phytec.com>