arm: dts: am572x-phycore-som: am572x-pcm-948: restructure ethernet PHY nodes
Restructuring the davinci_mdio node so that the PHYs
are slaves enables PHY drivers to get information
from the parent node.

The original method of using the davinci_mdio node
directly within the cpsw_emac nodes did not allow
for flexibility between the PHYs or to properly
add PHY fixups to individual PHYs.

Signed-off-by: Russell Robinson <rrobinson@phytec.com>