AM57SW-419 arm: dts: am57xx-pcm-*: only enable PWM backlight for AM572x
VIN1A_D0 is muxed as EHRPWM1A for PWM backlight
control, but this signal/PAD is not available
from the AM571x silicon so we are limiting PWM
and backlight support to the phyCORE-AM572x
configurations.

If the SOM is used with a custom carrier board,
another PAD/mux configuration can be implemented
for the same functionality and this patch reverted.

Signed-off-by: Russell Robinson <rrobinson@phytec.com>