AM57SW-607 arm: dts: am57xx-phycore-common: relocate CMEM from meta-phytec
The meta-phytec layer was creating the CMEM dtsi entries
for the appropriate MACHINE configurations, but this
results in different dtb configurations depending on if
the dtb was generated from the kernel source itself or
generated and modified by yocto.

Therefore, move the CMEM information back into the
appropriate kernel dtsi files so we can have consistent
builds regardless of how the source is accessed.

Signed-off-by: Russell Robinson <rrobinson@phytec.com>