net: ethernet: prueth: drop stale tx ptp message
A tx ptp event message is put in a wait queue by the driver
to wait for the tx timestamp.  If such a ptp message is still
waiting for the timestamp when the next message of the same
type is ready to be submitted to firmware, the old message
must have waited long enough for its timestamp. So instead
of just returning an error, the old message should be dropped
in order to make room in the wait queue for the new message.

Signed-off-by: WingMan Kwok <w-kwok2@ti.com>