net: ethernet: prueth: wait for cut-thru ptp timestamp
In HSR, when there is a heavy load of network traffic, it may take
a little longer for the HSR firmware to return the cut-through tx
timestamp for ptp event messages such as SYNC. If the cut-thru tx
timestamp is not available on first read, then instead of doing
nothing as in the implementation before this patch, the driver
should save the ptp message locally and schedule a delayed work to
try to read the timestamp again later. This patch makes such changes.

Signed-off-by: WingMan Kwok <w-kwok2@ti.com>