net: ethernet: prueth: Enable network storm prevention for EMAC
Storm prevention is supported via the DualEMAC firmware. This change
adds support for storm prevention via DualEMAC firmware to the PRUETH
driver. Storm prevention update was completed in a timer function that
also handled node table updates for HSR/PRP. The same function is
re-used (but renamed for clarity), and node table updates only enabled
if HSR/PRP is enabled.

Signed-off-by: Aaron Kramer <a-kramer@ti.com>