net: ethernet: prueth: Fix dual emac and hsr/prp co-existence issue
The rxq_size_defaults values had changed from 194 blocks to 206 blocks
when collision queue was eliminated from the driver/firmware at the
host interface. The idea at that time was to re-use the collision q
buffers in the normal queue. As there were 48 blocks used by collision q
48/4 = 12 blocks added to each of rx queue. However the collision q
buffers are located at the end of OCMC ram. So this extra blocks caused
BD descriptor area overlaps with other fields in shared/dram1 memory
and also OCMC ram extended to the reserved area. This manifest itself
as a loopback of frames from ingress to egress and causes kernel crash
when Dual EMAC is set up on one ICSS and HSR/PRP on another. So revert
the queue sizes to 194.

Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>