net: ethernet: prueth: introduce validation of queue sizes from DT
The queue sizes are configurable, however there is a limit of
1552 blocks available for Queues. Similar number of BDs are required
as well. Currently driver allows configure the queue sizes
at the ingress queues as well as egress queues. However if user
uses more buffers than actually supported by memory map, this can
cause malfunction of the driver. So introduce a validation scheme to
make sure the total number of blocks and BDs used are within 1552
limit set by the memory map. User is allowed to increase or decrease
the queue sizes within the above limit. If invalid values are used in
DT, restore the queue sizes to default values to ensure sane operation
and print a warning at the console as well.

To help in debug, update debugfs (rename stats to queue-info for this
purpose) to display the queue sizes used in the driver along with stats.

Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>