net: ethernet: prueth: add timeout to tx timestamp wait
Currently, an egress ptp event message is put in a wait queue by
the driver to wait for the tx timestamp. The waiting message
will then stay in the wait queue until either the tx timestamp
is available or the next message of the same type is ready to go
out. In the latter case, the waiting message is dropped to make
room for the new message. In the former case, the timestamp
returned is forwarded to user space application. However, if the
timestamp is not returned promptly, the timestamp forwarded is
obsolete and can be mis-used by user space application.  This
patch makes the changes such that if a timestamp is returned after
100ms, it is considered obsolete. Obsolete timestamp is not forwared
to user space and corresponding waiting message is then dropped.
100ms is determined experimentally in a network with a traffic
load of ~90M bps.

Signed-off-by: WingMan Kwok <w-kwok2@ti.com>