net: ethernet: prueth: Fix - fix debugfs initialization
Sometimes on boot an error message is seen saying that debugfs fails to
be created. This occurs because if PRUETH probe happens before filesystem
is up, probe succeeds even if creating debugfs fails. This change adds
a check so that if creating debugfs fails, then probe fails so that it
will try again and succeed once filesystem is up.

Signed-off-by: Aaron Kramer <a-kramer@ti.com>