net: ethernet: prueth: Fix - eth_type and boot params
Eth_type and other boot parameters were moved after the call to
prueth_netdev_init(), which caused the ports not to have the correct
features enabled (in particular, HSR/PRP offload was not being enabled
and all ports were brought up as EMAC type regardless of boot
parameter). This change moves all boot parameters back before
preuth_netdev_init() besides the mc_mac_mask because that needs the emac
struct which is initialized in that function.

Signed-off-by: Aaron Kramer <a-kramer@ti.com>