net: ethernet: prueth: Fix VLAN filter enable HSR/PRP
To adapt VLAN filtering to EMAC firmware a check was added to select
which PRU RAM (DATA0,DATA1,SHARED) to use depending on which firmware is
in use. The check in the function to enable VLAN filtering was missing
the HSR/PRP option (which uses SHARED RAM). This change adds the logic
and option to select SHARED RAM if HSR/PRP firmware is being used.

Signed-off-by: Aaron Kramer <a-kramer@ti.com>