AM57SW-655 arm: dts: am572x: add support for PCM-057-00001100I
This commit contains a workaround for CPSW EMAC support when only
cpsw_emac1 is connected to an ethernet PHY. Both cpsw_emac0 and
cpsw_emac1 must be enabled or the CPSW driver will fail to initialize
cpsw_emac1. Additionally, the 'dual_emac' property must be removed
from the 'mac' node or the CPSW driver will cause a kernel panic when
it attempts to initialize a NULL device.

Signed-off-by: Matt McKee <mmckee@phytec.com>