net: ethernet: prueth: Enable PTP PPS on AM3/4
This change enables PPS on AM3/4 for PTP via DualEMAC.

This change uses a different method to set the CMP1 register for SYNC
signal generation due to the 32-bit IEP wraparound. The CMP1 value must
be less than the IEP wraparound value (1 second), so it is truncated to
less than 1 second. Additionally, the SYNC signal must be reset *after*
IEP wraparound occurs or the SYNC pulse will trigger again because the
IEP value is still greater than CMP1. So SYNC is reset only during a
short interval immediately before the next SYNC pulse. This is an
incomplete fix as this will succeed the majority of the time but fail
when CMP1 is a very low value because the SYNC reset may happen before
IEP wraparound in that case.

Signed-off-by: Aaron Kramer <a-kramer@ti.com>