arm64: dts: k3-am6: move ti,psil-base and psil-config to icssgx node
The ti,psil-base property actually should reside in iccsgx node as
it defines the start psil thread for the icssg. This is used by the udma
device driver to allocate channel/flow id to communicate with the iccsg
pru/rtu. So move this to the iccsg node and use the iccsg phandle in the
channel spec of prueth device node. As a consequence, move the psil-config
as well to the same binding.

While at it also fix the compatibility string used for iccsg-prueth to
match with what is used in driver and dts.

Signed-off-by: Roger Quadros <rogerq@ti.com>
Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>
Tested-by: Murali Karicheri <m-karicheri2@ti.com>