dt-bindings: net: ti,icssg-prueth: update for interposer card
interposer card allows two ICSSG paired to provide a pair of
PRU Ethernet ports. Each PRU handles a egress or ingress path
in this configuration. A new compatibility is defined to identify
this particular hardware configuration where RGMII signals of
data path of the ICSSG1 gets connected to the same pair of
Phys that is connected to ICSSG0. This is used to customize the
driver. mii-gt-rt-paired added to points to the mii-g-rt of ICSSG1.

Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>