dt-bindings: net: ti,icssg-prueth: bindings update for 100M support
PRU Ethernet ports should support both 100M/1G operation. To support
this, the MII interface parameters need to be configured for proper
operation at these link speeds. This requires the mii-rt and mii-rt-paired
syscon regmap base address to be available to the driver to configure
the TX IPG values differently for 100M and 1G link speeds. So add the
bindings details to the documentation.

Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>