net: ethernet: ti: iep: fix the PPS glitch
This patch fixes the IEP-based PPS glitch that an invalid PPS pluse is
generated right after the previous one. The root cause of this problem is
that the IEP counter is read right before the PPS match time and then the
invalid time difference is used to adjust the next CMP value by mistake.

Signed-off-by: Eric Ruei <e-ruei1@ti.com>