net: ti: icssg_prueth: Add support for 100M link operation
This adds support for 100M link over PRU ports. The phy auto
negotiates and update the link speed to the driver. When 100M link
is used, Gig enable bit in RGMII CFG register is reset. Also use an
IPG value of 0x166 for 100M link and 0x18 for 1G link. While at it
also fix an error check for miig_rt for ICSSG0.

Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>